Make experience and gold long integer
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1122 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/umoria.d/types.h
@@ -77,9 +77,9 @@ typedef struct player_type
       vtype sex;	/* Sex of character	*/
       vtype title;	/* Character's title	*/
       vtype tclass;	/* Character's class	*/
-      int max_exp;	/* Max experience*/
-      int exp;	        /* Cur experience	*/
-      int au;	        /* Gold		*/
+      long int max_exp;	/* Max experience*/
+      long int exp;	/* Cur experience	*/
+      long int au;	/* Gold		*/
       wordint age;	/* Characters age*/
       wordint ht;	/* Height	*/
       wordint wt;	/* Weight	*/
